requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
numpy==1.26.4
scikit-learn==1.5.2
pydantic==2.9.2
python-dotenv==1.0.1
sentence-transformers==3.3.1
pinecone-client==5.0.1
tf-keras==2.20.1

# Database & Caching
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis==5.2.1
alembic==1.14.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-asyncio==0.24.0
httpx==0.27.2

# Code Quality
flake8==7.1.1
black==24.10.0
PyYAML==6.0.2
